fix(check_gitlab_pull_mirror): use correct timezone handling
the old code used a fixed offset of 2 hours from the local time
(probably CEST) to compare with the timestamp from the gitlab API which
is in UTC ("Z" at the end of the timestamp)

The updated code should handle this correctly - internal comparision is
in UTC and the text output is in the local timezone.

I also added some exception handling and error out I needed to analyse
some issues I had with the check.